Fix GSNSignatureBouncer signature bug
The GSNSignatureBouncer currently accepts a zero address as a trusted signer. Even more unexpectedly however, is the failure to initialize the GSNSignatureBouncer: In this case the bouncer will accept no signature at all, as _getTrustedSigner() will resolve a zero address for unset values. For people who deploy and fail to initialize their bouncer, their GSN contract will simply "magically" work for anyone who does not provide a signature, but they won't know that because their valid provided signatures will always fail.

I've added a catch so that the trustedSigner can not be set to 0x0, and a check so that it will not relay transactions when no `trustedSigner` has been set at all.